+++ examples/mock.rs
@@ -26,23 +26,6 @@ async fn make_insert(client: &Client, data: &[SomeRow]) -> Result<()> {
     insert.end().await
 }
 
-#[cfg(feature = "watch")]
-async fn make_watch(client: &Client) -> Result<(u64, SomeRow)> {
-    client
-        .watch("SELECT max(no) no FROM test")
-        .fetch_one::<SomeRow>()
-        .await
-}
-
-#[cfg(feature = "watch")]
-async fn make_watch_only_events(client: &Client) -> Result<u64> {
-    client
-        .watch("SELECT max(no) no FROM test")
-        .only_events()
-        .fetch_one()
-        .await
-}
-
 #[tokio::main]
 async fn main() {
     let mock = test::Mock::new();
@@ -69,23 +52,4 @@ async fn main() {
     make_insert(&client, &list).await.unwrap();
     let rows: Vec<SomeRow> = recording.collect().await;
     assert_eq!(rows, list);
-
-    // How to test WATCH.
-    #[cfg(feature = "watch")]
-    {
-        // Check `CREATE LIVE VIEW` (for `watch(query)` case only).
-        let recording = mock.add(test::handlers::record_ddl());
-        mock.add(test::handlers::watch(list.into_iter().map(|row| (42, row))));
-        let (version, row) = make_watch(&client).await.unwrap();
-        assert!(recording.query().await.contains("CREATE LIVE VIEW"));
-        assert_eq!(version, 42);
-        assert_eq!(row, SomeRow { no: 1 });
-
-        // `EVENTS`.
-        let recording = mock.add(test::handlers::record_ddl());
-        mock.add(test::handlers::watch_only_events(3..5));
-        let version = make_watch_only_events(&client).await.unwrap();
-        assert!(recording.query().await.contains("CREATE LIVE VIEW"));
-        assert_eq!(version, 3);
-    }
 }